移出时态
